Applying fixes and modifying certain things for SLAMBench V1.2.
The main fixes are hard coding a 3 into the sycl/interface.h depth read function. Rather than using a sizeof() call, as SYCL sizes are slightly different than the standard sizes used by SLAMBench. This size difference was causing some lost frames and an garbage data message to be printed.

The second fix also appears to be related to the SYCL size differences. When casting an array of floats to a float3 for easier assignment it was emitting a segmentation fault. The current work around to this is to assign it less elegantly, but safer (using element by element assignment, so possibly a little slower and less elegant).

Signed-off-by: Andrew Gozillon <andrew.gozillon@uws.ac.uk>