Modifying the readme.md sycl section to include the DAGR repository and adding further information to make it easier to compile the SYCL modules.
I also modified a static_assert in the sycl/kernels.cpp as it was causing problems with ComputeCpp version 0.3.2 (and possibly above). The problem was a little odd, It is possible that in ComputeCpp 0.3.2 a compiler bug was introduced that for some reason is running the static_assert twice and looking for one with and one without an address space appended to the type. Perhaps something to do with this change: "Improved error message for address-space resolution". Another possibility is that this template gets instantiated twice in ComputeCpp 0.3.2 with different types. Either way the static_assert was failing as it was getting passed twice with different types each time, one of which was incorrect. The pass has been modified to check for both types (as it functions fine in either case) this allows compilation of the SYCL modules for ComputeCpp versions 0.2.1 to the current latest 0.5.0.