Fix Build System (requires newest SYCLCC) and modify documentation
I've updated the documentation to be more up-to-date, especially after
the recent script updates.

I've made an attempt at fixing the build system when CXX is set to
syclcc. You should now be able to compile all modules when it's set.

The main change was modfiyng the -DSYCL compiler definition to only be
speciifed for the SYCL module.

There was another modification required that links to some other syclcc
modifications that had to be done. When using syclcc (and by default
ComputeCpp) as the main compiler, it unfortunately doesn't come packaged
with OpenMP libraries which means the OpenMP module will fail unless a
user has self installed the OpenMP libraries.

The fix for this requried the addition of a new "driver" argument for
syclcc and the application of this argument for the OpenMP module when
syclcc is in use (-syclcc-use-host). Which essentially says use the host
compiler, which by default is g++. This means it'll compile smoothly as
g++ contains omp headers and libraries.